move OcrMenu component out of Geral subfolder
